refactor v2 validator dependency tests
